Accounting for time
Several of these items needed to be shifted to past tense.

Do we want to go all the way back to 2015? I think this list should be RECENT news. Anything significant that is more than a few months old (maybe we can 6 to 9 months max) should be moved to the timeline. Nothing from August is still "News"