More changes to reflect move to Linux Foundation
Added the new legal name. Also removed the line about "people can contribute whether they join or not." This is one of several places where the policy that people must join the project to contribute to the documents should be stipulated. But since we don't have the mechanism to join in place just yet, I didn't want to write in the stipulation at this point.