First round of changes to this
I went in to fix something minor and found this page still full of references to the Uptane Alliance (an organization that no longer exists). I'm opening this PR to begin getting this corrected but I have two questions:

I have put my name in several places in the website, but hesitate to do so where Sebastien's name appears here, as these would be requests/questions of a technical nature. Should perhaps Marina be listed here? Or should I insert my name with the idea that I will refer it on to someone else to address.