Removing Time Server
While we scoured the document pages, it seems we did not check the website for mentions of time servers. I made one catch on this page and did a "quick and dirty fix" on the diagram, which should be re-drawn at some point. Please check to make sure this is serviceable, at least till we do a more complete re-design.