Adding conference announcement to home page
This is the only way I could figure out to post something about the meeting on the home page. I have created a PDF of the "save the date" flyer and stored it with the publications on this site, which I will add to this pull request. If someone can think of a better way to spotlight the workshop on this page, let me know.